Add files via upload
The Word document that is labeled "Professor's Instructions..." reflects the design and expectations of the created project.

The Word document that is labeled "testCases..." reflects the implementation and execution of different methods and codes within the main method of the .java file associated with this project. Along with the executed code, there is present the visual representation of the effect of the code that is executed on the console output that can be observed in this word document.